+++ app.config.ts
@@ -1,7 +1,7 @@
 export default defineAppConfig({
   github : 'https://github.com/kkan0615/blog-nuxt',
   blogCategories: [
-    'example',
+    // 'example',
     'programming',
   ],
   blogTags: {